requirements.txt

flask
flask-cors
easyocr
opencv-python-headless
Pillow
torch==2.9.0
torchvision==0.24.0
torchaudio==2.9.0+cpu
